+++ README.md
@@ -5,9 +5,9 @@ Code for Atmega16/32 based microcontrollers
 # Libraries
 
 1. Serial Communication library
-2. Keypad Interfacing library
-3. LCD library
-4. I2C/TWI Communication library
+2. I2C/TWI Communication library
+3. Keypad Interfacing library
+4. LCD Interfacing library
 5. DS1307 Interfacing library
 
 # To Add
